Module hdkeygen::account

account based wallet, does not really have any Hierarchical element in it thought it is part of the keygen entity so adding it in

On the chain, an account can group stake while a rindex or a bip44 cannot as they represent individual coins, which have stake but they are not grouped and cannot be be controlled without having an account to group them

Structs

Account

AccountId

Type Definitions

Seed