Struct hdkeygen::Key

pub struct Key<K, P> {
    key: K,
    path: DerivationPath<P>,
    derivation_scheme: DerivationScheme,
}

convenient wrapper around the Key.

Fields
key: Kpath: DerivationPath<P>derivation_scheme: DerivationScheme
Implementations

impl Key<XPrv, Bip44<Root>>

pub fn purpose(&self, derivation: HardDerivation) -> Key<XPrv, Bip44<Purpose>>

pub fn bip44(&self) -> Key<XPrv, Bip44<Purpose>>

pub fn chimeric_bip44(&self) -> Key<XPrv, Bip44<Purpose>>

impl Key<XPrv, Bip44<Purpose>>

pub fn coin_type( &self, derivation: HardDerivation ) -> Key<XPrv, Bip44<CoinType>>

pub fn cardano(&self) -> Key<XPrv, Bip44<CoinType>>

impl Key<XPrv, Bip44<CoinType>>

pub fn account(&self, derivation: HardDerivation) -> Key<XPrv, Bip44<Account>>

impl<K> Key<K, Bip44<Account>>

const EXTERNAL: SoftDerivation = DerivationPath<Bip44<bip44::Account>>::EXTERNAL

const INTERNAL: SoftDerivation = DerivationPath<Bip44<bip44::Account>>::INTERNAL

const ACCOUNT: SoftDerivation = DerivationPath<Bip44<bip44::Account>>::ACCOUNT

pub fn id(&self) -> HardDerivation

impl Key<XPrv, Bip44<Account>>

pub fn change(&self, derivation: SoftDerivation) -> Key<XPrv, Bip44<Change>>

pub fn external(&self) -> Key<XPrv, Bip44<Change>>

pub fn internal(&self) -> Key<XPrv, Bip44<Change>>

pub fn account(&self) -> Key<XPrv, Bip44<Change>>

impl Key<XPub, Bip44<Account>>

pub fn change(&self, derivation: SoftDerivation) -> Key<XPub, Bip44<Change>>

pub fn external(&self) -> Key<XPub, Bip44<Change>>

pub fn internal(&self) -> Key<XPub, Bip44<Change>>

pub fn account(&self) -> Key<XPub, Bip44<Change>>

impl Key<XPrv, Bip44<Change>>

pub fn address(&self, derivation: SoftDerivation) -> Key<XPrv, Bip44<Address>>

impl Key<XPub, Bip44<Change>>

pub fn address(&self, derivation: SoftDerivation) -> Key<XPub, Bip44<Address>>

pub fn addresses( &self, range: SoftDerivationRange ) -> KeyRange<'_, XPub, SoftDerivationRange, Bip44<Change>, Bip44<Address>>

impl Key<XPub, Bip44<Address>>

pub fn address_single(&self, discrimination: Discrimination) -> Address

pub fn address_account(&self, discrimination: Discrimination) -> Address

pub fn address_group( &self, discrimination: Discrimination, group: PublicKey<Ed25519> ) -> Address

impl<K, P> Key<K, P>

pub fn new_unchecked( key: K, path: DerivationPath<P>, derivation_scheme: DerivationScheme ) -> Self

create a Key with the given component

does not guarantee that the derivation path is actually the one that lead to this key derivation.

pub fn path(&self) -> &DerivationPath<P>

get the derivation path that lead to this key

pub fn coerce_unchecked<Q>(self) -> Key<K, Q>

impl<P> Key<XPrv, P>

pub fn public(&self) -> Key<XPub, P>

retrieve the associated public key of the given private key

pub fn sign<T, B>(&self, message: B) -> Signature<T>where B: AsRef<[u8]>,

create a signature for the given message and associate the given type T to the signature type.

pub fn verify<T, B>(&self, message: B, signature: &Signature<T>) -> boolwhere B: AsRef<[u8]>,

verify the signature with the private key for the given message

pub fn chain_code(&self) -> [u8; 32]

get key’s chain code

pub(crate) fn derive_unchecked<Q>(&self, derivation: Derivation) -> Key<XPrv, Q>

derive the private key against the given derivation index and scheme

pub(crate) fn derive_path_unchecked<'a, Q, I>( &'a self, derivation_path: I ) -> Key<XPrv, Q>where I: IntoIterator<Item = &'a Derivation>,

derive the private key against the given derivation index and scheme

impl<P> Key<XPub, P>

pub fn verify<T, B>(&self, message: B, signature: &Signature<T>) -> boolwhere B: AsRef<[u8]>,

verify the signature with the public key for the given message

pub fn public_key_slice(&self) -> &[u8]

get the public key content without revealing the chaincode.

pub fn public_key(&self) -> &XPub

pub fn pk(&self) -> PublicKey<Ed25519>

pub(crate) fn derive_unchecked<Q>( &self, derivation: SoftDerivation ) -> Key<XPub, Q>

derive the private key against the given derivation index and scheme

impl<P> Key<SecretKey<Ed25519Extended>, P>

pub fn sign<T, B>(&self, message: B) -> Signature<T>where B: AsRef<[u8]>,

create a signature for the given message and associate the given type T to the signature type.

pub fn pk(&self) -> PublicKey<Ed25519>

impl Key<XPrv, Rindex<Root>>

pub fn key( &self, derivation_path: &DerivationPath<Rindex<Address>> ) -> Key<XPrv, Rindex<Address>>

pub fn hd_key(&self) -> HdKey

get an address recovering object, this object can be used to check the ownership of addresses